lifted up state and finished deals page
