+++ src/components/pages/Home.js
@@ -1,10 +1,12 @@
 import React from 'react';
 import Items from '../items/Items';
 
-const Home = () => {
+const Home = (props) => {
+  const { items } = props;
+
   return (
     <main className='main'>
-      <Items />
+      <Items items={items} />
     </main>
   );
 };
