+++ src/css/App.css
@@ -12,6 +12,12 @@ nav {
   width: 18rem;
 }
 
+.apology {
+  margin-top: 3rem;
+  margin-left: auto;
+  grid-column: 1/3;
+}
+
 .card-img-top {
   height: 50%;
   padding: 1rem;
